Move `Viz` layer to new `.ui` mod
